+++ src/tno/euphorie/__init__.py
@@ -2,4 +2,4 @@ from zope.i18nmessageid import MessageFactory as mf
 MessageFactory = mf("tno.euphorie")
 del mf
 
-__all__ = [ "MessageFactory" ]
+__all__ = ["MessageFactory"]