+++ src/tno/euphorie/tests/test_api.py
@@ -1,4 +1,37 @@
+import unittest
 from tno.euphorie.testing import TnoEuphorieFunctionalTestCase
+
+
+class get_json_decimal_tests(unittest.TestCase):
+    def get_json_decimal(self, *a, **kw):
+        from ..api import get_json_decimal
+        return get_json_decimal(*a, **kw)
+
+    def test_missing_required_field(self):
+        self.assertRaises(KeyError,
+                self.get_json_decimal, {}, 'field', required=True)
+
+    def test_missing_optional_field(self):
+        self.assertEqual(
+                self.get_json_decimal({}, 'field', required=False,
+                    default='default'),
+                'default')
+
+    def test_bad_type(self):
+        self.assertRaises(ValueError,
+                self.get_json_decimal, {'field': 'dummy'}, 'field')
+
+    def test_proper_value(self):
+        import decimal
+        self.assertEqual(
+                self.get_json_decimal({'field': 15}, 'field'),
+                decimal.Decimal('15'))
+
+    def test_rounding(self):
+        import decimal
+        self.assertEqual(
+                self.get_json_decimal({'field': 25.6}, 'field'),
+                decimal.Decimal('25.6'))
 
 
 class ViewBrowserTests(TnoEuphorieFunctionalTestCase):
@@ -47,6 +80,7 @@ class CompanyTests(TnoEuphorieFunctionalTestCase):
         return SurveySession(dutch_company=company)
 
     def test_do_GET_result(self):
+        import json
         from zope.publisher.browser import TestRequest
         context = self.create_context()
         view = self.Company(context, TestRequest())
@@ -60,13 +94,27 @@ class CompanyTests(TnoEuphorieFunctionalTestCase):
                      'absentee-percentage', 'accidents',
                      'incapacitated-workers', 'arbo-expert',
                      'works-council-approval']))
+        json.dumps(response)  # Check it is serializable
         self.assertEqual(response['type'], 'dutch-company')
         self.assertEqual(response['title'], u'Acme B.V.')
+        self.assertEqual(response['absentee-percentage'], None)
         self.assertEqual(response['employees'], '40h')
         self.assertEqual(response['submitted'], '2012-06-06')
 
+    def test_do_GET_with_absentee_percentage(self):
+        import decimal
+        import json
+        from zope.publisher.browser import TestRequest
+        context = self.create_context()
+        context.dutch_company.absentee_percentage = decimal.Decimal('12.34')
+        view = self.Company(context, TestRequest())
+        response = view.do_GET()
+        json.dumps(response)  # Check it is serializable
+        self.assertEqual(response['absentee-percentage'], 12.34)
+
     def test_do_PUT(self):
         import datetime
+        import decimal
         from zope.publisher.browser import TestRequest
         context = self.create_context()
         view = self.Company(context, TestRequest())
@@ -75,10 +123,12 @@ class CompanyTests(TnoEuphorieFunctionalTestCase):
                           'city': u'Ons Dorp'},
                       'department': u'Incasso',
                       'accidents': 15,
+                      'absentee-percentage': 25.6,
                       'employees': 'over25',
                       'works-council-approval': '2012-06-04',
                      }
-        view.do_PUT()
+        response = view.do_PUT()
+        self.assertEqual(response['type'], 'dutch-company')
         self.assertEqual(
                 context.dutch_company.address_visit_address,
                 u'Dorpsstraat 15')
@@ -87,6 +137,9 @@ class CompanyTests(TnoEuphorieFunctionalTestCase):
                 u'Ons Dorp')
         self.assertEqual(context.dutch_company.department, u'Incasso')
         self.assertEqual(context.dutch_company.accidents, 15)
+        self.assertEqual(
+                context.dutch_company.absentee_percentage,
+                decimal.Decimal('25.6'))
         self.assertEqual(context.dutch_company.employees, 'over25')
         self.assertEqual(
                 context.dutch_company.works_council_approval,